feat: Add sync and async endpoint

+++ src/main/java/com/marcos/async/web/dto/AsyncMetricsDto.java
@@ -1,8 +1,6 @@
 package com.marcos.async.web.dto;
 
-import java.math.BigInteger;
-
 public record AsyncMetricsDto(
     Integer requestsCount,
-    BigInteger durationInMilliseconds
+    Long durationInMilliseconds
 ) {}
